requirements.txt
absl-py==1.2.0
adal==1.2.7
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
applicationinsights==0.11.10
argcomplete==1.12.3
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
azure-common==1.1.28
azure-core==1.21.1
azure-graphrbac==0.61.1
azure-identity==1.7.0
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==20.1.0
azure-mgmt-storage==19.1.0
azureml==0.2.7
azureml-automl-core==1.38.1
azureml-core==1.38.0.post2
azureml-dataprep==2.26.0
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.2.0
azureml-dataset-runtime==1.38.0
azureml-pipeline==1.38.0
azureml-pipeline-core==1.38.0
azureml-pipeline-steps==1.38.0
azureml-sdk==1.38.0
azureml-telemetry==1.38.0.post1
azureml-train-automl-client==1.38.0.post3
azureml-train-core==1.38.0
azureml-train-restclients-hyperdrive==1.38.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
cloudpickle==2.1.0
colorama==0.4.5
contextlib2==21.6.0
cryptography==36.0.2
dbus-python==1.2.16
decorator==5.1.1
distro==1.4.0
docker==5.0.3
dotnetcore2==2.1.23
filelock==3.8.0
fire==0.4.0
flatbuffers==1.12
frozenlist==1.3.1
fsspec==2022.8.2
fusepy==3.0.1
future==0.18.2
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
humanfriendly==10.0
hyperopt==0.1.2
idna==3.4
importlib-metadata==4.12.0
isodate==0.6.1
jeepney==0.8.0
jmespath==0.10.0
joblib==1.1.0
json-tricks==3.15.5
jsonpickle==2.2.0
keras==2.9.0
Keras-Preprocessing==1.1.2
knack==0.8.2
libclang==14.0.6
lightgbm==3.3.2
lightning-utilities==0.3.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
msal==1.18.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
multidict==6.0.2
ndg-httpsclient==0.5.1
networkx==2.8.6
-e git+https://github.com/microsoft/nni.git@128bc7ce9b174ecd90c0fbb720cb3c9fc8e41370#egg=nni
numpy==1.22.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.6.0.163
nvidia-ml-py==11.515.75
nvidia-tensorrt==8.4.1.5
oauthlib==3.2.0
onnx==1.12.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
paramiko==2.11.0
pathspec==0.10.1
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.4
portalocker==2.5.1
prettytable==3.4.1
protobuf==3.19.4
psutil==5.9.2
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.13.0
PyGObject==3.36.0
PyJWT==2.4.0
pymongo==4.2.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PythonWebHDFS==0.2.3
pytools==2022.1.13
pytorch-lightning==1.8.1
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.21.0
rsa==4.9
schema==0.7.5
scikit-learn==1.0.2
scipy==1.8.0
SecretStorage==3.3.3
simplejson==3.17.6
six==1.14.0
ssh-import-id==5.10
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchmetrics==0.9.3
torchvision==0.11.3+cu113
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.3.0
urllib3==1.26.13
wcwidth==0.2.5
websocket-client==1.4.1
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
